refactor: split the mirror package by responsibility
mirror.go had grown to a thousand lines mixing routing, branch pinning,
tree rewriting, the ingestion task, upstream fetching, and resolve
serving. Each concern now lives in its own file, branch persistence
moved from index.go next to the resolution logic it serves, and the
default proxy construction became newUpstreamProxy. Pure code motion,
no behavior change.
